Salvar funções e menus extras do profissional em andamento.

+++ sishosp/src/br/gov/al/maceio/sishosp/acl/dao/MenuDAO.java
@@ -906,7 +906,7 @@ public class MenuDAO {
 		return lista;
 	}
 
-	public ArrayList<Menu> listarMenuItemTargetEditUser(Integer idUsuario)
+	public ArrayList<Menu> listarMenuItemTargetEditUser(Long idUsuario)
 			throws ProjetoException {
 
 		String sql = "select me.id, me.descricao, me.codigo, me.indice, me.tipo, "
@@ -923,7 +923,7 @@ public class MenuDAO {
 		try {
 			conexao = ConnectionFactory.getConnection();
 			PreparedStatement stmt = conexao.prepareStatement(sql);
-			stmt.setInt(1, idUsuario);
+			stmt.setLong(1, idUsuario);
 			ResultSet rs = stmt.executeQuery();
 
 			while (rs.next()) {
@@ -964,7 +964,7 @@ public class MenuDAO {
 		return lista;
 	}
 
-	public ArrayList<Menu> listarMenusPerfil(Integer idPerfil)
+	public ArrayList<Menu> listarMenusPerfil(Long idPerfil)
 			throws ProjetoException {
 
 		String sql = "select me.id, me.descricao, me.codigo, me.indice, me.tipo, "
@@ -981,7 +981,7 @@ public class MenuDAO {
 		try {
 			conexao = ConnectionFactory.getConnection();
 			PreparedStatement stmt = conexao.prepareStatement(sql);
-			stmt.setInt(1, idPerfil);
+			stmt.setLong(1, idPerfil);
 			ResultSet rs = stmt.executeQuery();
 
 			while (rs.next()) {
